Add files via upload
This program will allow a user to find the overlapping region of two overlapping pairs. The program takes in two genes that overlap in fasta format, finds the longest shared sequence and then prints the results to stdout. Both of the input genes will be printed to stdout with the overlapping region in a different text color so the user may identify it. The overlapping sequence will also be printed.